+++ CHANGELOG.md
@@ -18,6 +18,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v8.6.0 - 2020-08-19
+
+### Added
+
+- 決済承認アクションのinstrumentを決済サービスとして定義
+
+### Changed
+
+- 決済承認アクションのobject.typeOfを'Payment'に統一
+
 ## v8.5.0 - 2020-08-19
 
 ### Added